Correct Opal regexp inner working to increase correctness.
This patch cherry-picks updated version of opal/opal#2610. This
update cleaned up Opal's regexp implementation quite a lot to fix
opal/opal#2616 which in turn caused that we got 2 more tests passed
on Plurimath and only the obvious failures remain on Parslet (that
is after re-enabling some previously disabled tests, most of which
pass now). This patch also removes the fix from #22, as it's not
needed anymore.