Fix all the remaining test cases. This fixes #8.
This pull request takes an updated opal/opal#2610 - there was an
issue in a previous version that treated `/[^"]/` as a multiline
regexp which caused Parslet to produce invalid results. The other
cases of this kind were fixed before.

In addition this pull request cherry-picks opal/opal#2620 which
fixes a mistake when trying to use `String#chars` on a String
containing over-the-BMP characters, so characters with Unicode
codepoints > 0xffff were handled incorrectly. While those don't
currently happen in tests other than our unit test for an XML
engine abstraction, having a support for those may be crucial
in further development.